Q5 Please test using your fav IDE.
Write an application that displays a series of at least 10 student ID numbers (that you have

stored in an array) and asks the user to enter a test letter grade for thestudent. Create

anExceptionclass namedGradeExceptionthat contains astatic publicarray of valid grade

letters (‘A’, ‘B’, ‘C’, ‘D’, ‘F’,and‘I’),whichyoucanuseto determine whether a grade entered

from the application is valid. In your application, throw aGradeExceptionif the user does not

enter a valid letter grade. Catch the GradeExceptionand then display an appropriate message.

In addition, store an‘I’ (for Incomplete) for any student for whom an exception is caught. At

the end of the application, display all the student IDs and grades.